requirements.txt
attrdict==2.0.1
cached-property==1.5.1
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Deprecated==1.2.10
facebook-wda==1.1.10
idna==2.10
Pillow==7.2.0
py==1.9.0
requests==2.24.0
retry==0.9.2
six==1.15.0
urllib3==1.25.10
wrapt==1.12.1
